refactor: delete experience folder
