+++ src/components/back-button/styled/icon.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import BackIcon from '../../../icons/back';
+import BackIcon from '../../icons/back';
 
 const StyledIcon = styled(BackIcon)`
     margin-right: 8px;
