+++ src/components/initializing-app/index.js
