+++ src/components/logo-link/styled/image.js
