+++ src/components/message/index.js
@@ -4,9 +4,9 @@ import { bindActionCreators } from 'redux';
 import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import ButtonIcon from 'react-rainbow-components/components/ButtonIcon';
-import CloseIcon from '../../icons/close';
+import CloseIcon from '../icons/close';
 import MessageIcon from './icon';
-import hideMessage from '../../../redux/actions/app/hide-message';
+import hideMessage from '../../redux/actions/app/hide-message';
 import './styles.css';
 
 function Message(props) {
