+++ src/components/social-login/icons/facebookIcon.js
