+++ src/components/social-login/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { FormattedMessage } from 'react-intl';
-import { loginWithFacebook, loginWithGoogle } from '../../../redux/actions/authentication';
+import { loginWithFacebook, loginWithGoogle } from '../../redux/actions/authentication';
 import StyledArticle from './styled/article';
 import StyledLoginButton from './styled/loginButton';
 import StyledFacebookButton from './styled/facebookButton';
