+++ src/components/social-login/styled/icon.js
