+++ src/components/termsConditions-and-privacyPolicy/idnex.js
