+++ src/components/termsConditions-and-privacyPolicy/styles.css
