+++ src/pages/public/terms-condition/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Prismic, QueryAt } from 'react-prismic-cms';
-import Content from '../../../components/experiences/prismic-content';
-import BackButton from '../../../components/experiences/back-button';
+import Content from '../../../components/prismic-content';
+import BackButton from '../../../components/back-button';
 import './styles.css';
 
 function TermsAndConditions(props) {
