+++ src/routes.js
@@ -6,8 +6,8 @@ import {
 } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import history from './history';
-import AuthRoute from './components/experiences/auth-route';
-import PublicRoute from './components/experiences/public-route';
+import AuthRoute from './components/auth-route';
+import PublicRoute from './components/public-route';
 import PublicRoutes from './pages/public';
 import AuthRoutes from './pages/auth';
